[RESTRUCTURE] 동일한 클래스 타입들 모두 패키지로 묶음
