+++ src/main/java/com/gamja/tiggle/reservation/application/port/in/GetSectionUseCase.java
@@ -1,4 +1,4 @@
-package com.gamja.tiggle.reservation.application.in;
+package com.gamja.tiggle.reservation.application.port.in;
 
 
 import com.gamja.tiggle.common.BaseException;
