+++ src/main/java/com/gamja/tiggle/reservation/application/port/in/GetTimesUseCase.java
@@ -1,4 +1,4 @@
-package com.gamja.tiggle.reservation.application.in;
+package com.gamja.tiggle.reservation.application.port.in;
 
 import com.gamja.tiggle.common.BaseException;
 import com.gamja.tiggle.reservation.domain.Times;
